% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atm_pressure.R
\name{atm_pressure}
\alias{atm_pressure}
\title{Atmospheric pressure}
\usage{
atm_pressure(z)
}
\arguments{
\item{z}{Elevation above sea level in meters.}
}
\value{
Atmospheric pressure in kPa.
}
\description{
This function is used to estimate the The atmospheric pressure, P, by using a simplification of the ideal gas law, assuming 20°C for a standard atmosphere.
Please, see the equation 7 from the FAO Irrigation and drainage paper 56.
}
\examples{
atm_pressure(z= 1120)

}
\references{
Allen et al., 1998. Crop Evapotranspiration. Guidelines for computing crop water requirements. FAO Irrigation and drainage paper 56.
\url{http://www.fao.org/3/X0490E/x0490e00.htm}.
}
